upd chatId field type (int -> Long)

+++ bot/src/main/java/edu/java/bot/entities/CommandCallContext.java
@@ -8,7 +8,7 @@ import java.util.List;
 @AllArgsConstructor
 @Getter public class CommandCallContext {
     @NotNull private User user;
-    private int chatId;
+    private Long chatId;
     @NotNull private String command;
     private List<String> commandArgs;
 }
